requirements.txt
numpy==1.18.5
scipy==1.5.0
flask==1.1.2
pylint==2.5.3
imageio==2.9.0
google-cloud-storage==1.30.0
requests==2.24.0

